% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{forecastExtract}
\alias{forecastExtract}
\title{Extracting values from forecast objects using OrniStatsData for use in ggplot2}
\usage{
forecastExtract(fcast)
}
\arguments{
\item{fcast}{Object of type `forecast` from which the values are extracted}
}
\value{
Data frame containing the point forecast, 50 and 80 percent confidence interval values
}
\description{
Extracting values from forecast objects using OrniStatsData for use in ggplot2
}
