% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{recordTable}
\alias{recordTable}
\title{Generate table of records of species in OrniStats data.}
\usage{
recordTable(qualifier, path = NA)
}
\arguments{
\item{path}{File path to save the table. Default value is NA where tables are not saved.}

\item{qualifer}{Single or list of quoted 5-letter abbreviation(s) for the species whose records are being tabulated. To check which qualifiers you need, please consult the list in the GitHub repository for this package.}
}
\value{
The OrniStats table for the species inputted.
}
\description{
Generate table of records of species in OrniStats data.
}
