% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{speciesPlots}
\alias{speciesPlots}
\title{Generate the graphs for length of stay, arrival date and arrival location of species in OrniStats data.}
\usage{
speciesPlots(qualifier, path = NA)
}
\arguments{
\item{path}{File path to save the figure. Default value is `NA` where graphs are not saved.}

\item{qualifer}{Single or list of quoted 5-letter abbreviation(s) for the species whose records are being analysed. To find out the qualifier for a species, please consult the list in the GitHub repository for this package.}
}
\value{
ggplot figures summarising length of stay, arrival date and arrival location trends for the inputted species
}
\description{
Generate the graphs for length of stay, arrival date and arrival location of species in OrniStats data.
}
