% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{speciesRegression}
\alias{speciesRegression}
\title{Conduct simple regression analysis to compare associations between variables in the OrniStats data.}
\usage{
speciesRegression(qualifier, x, y, max_degree = 10, approx_dates = FALSE)
}
\arguments{
\item{qualifier}{Single or list of quoted 5-letter abbreviation(s) for the species whose records are being analysed. To find out the qualifier for a species, please consult the list in the GitHub repository for this package.}

\item{x, y}{Variables being analysed by regression. The inputs should correspond to the name of the variable column in the `OrniStatsData` or `OrniStatsLocations` objects.}

\item{max_degree}{Maximum degree polynomial at which regression analysis is conducted. If program is running slowly, set this value to 5.}

\item{approx_dates}{Remove records where the date of arrival or departure is approximate. Defaults to `FALSE`, but should be set to `TRUE` if either `x` or `y` is the date of arrival or departure at below an annual level.}
}
\value{
Summary table of the model selected in the readline.
}
\description{
Conduct simple regression analysis to compare associations between variables in the OrniStats data.
}
