% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{speciesStats}
\alias{speciesStats}
\title{Generate summary statistics for species or species groups in OrniStats data.}
\usage{
speciesStats(qualifier)
}
\arguments{
\item{qualifier}{Single or list of quoted 5-letter abbreviation(s) for the species whose records are being analysed. To find out the qualifier for a species, please consult the list in the GitHub repository for this package.}
}
\value{
Output data frame containing measures of variability, heterogeneity, fidelity and mortality for given species or species groups.
}
\description{
Generate summary statistics for species or species groups in OrniStats data.
}
