% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Master Script.R
\name{yearlyForecast}
\alias{yearlyForecast}
\title{10-year forecasts (ETS or ARIMA) for yearly frequencies of bird species in OrniStats data.}
\usage{
yearlyForecast(qualifier, path = NA)
}
\arguments{
\item{path}{File path to save the figure. Default value is NA where graphs are not saved.}

\item{qualifer}{Single or list of quoted 5-letter abbreviation(s) for the species whose records are being analysed. To find out the qualifier for a species, please consult the list in the GitHub repository for this package.}
}
\value{
ggplot figure with historical annual frequencies, 10-year point forecast of chosen type in readline, and 50 and 80 percent confidence interval values.
}
\description{
10-year forecasts (ETS or ARIMA) for yearly frequencies of bird species in OrniStats data.
}
